+++ src/extension.ts
@@ -36,7 +36,25 @@ async function initVs() {
 	vscode.window.showInformationMessage(`Created a new file: ${filePath.toString()}`);
 }
 
+function checkOnStartup() {
+	function error(message: string) {
+		vscode.window.showErrorMessage(`Error during Tarantool plugin initialization: ${message}. Try reinstalling the extension from the marketplace`);
+		return false;
+	}
+
+	const brokenPaths = annotationsPaths.filter((path) => !fs.existsSync(path));
+	if (brokenPaths.length > 0) {
+		return error(`type annotations are not available at ${brokenPaths.join(', ')}`);
+	}
+
+	return true;
+}
+
 export function activate(context: vscode.ExtensionContext) {
+	if (!checkOnStartup()) {
+		return;
+	}
+
 	const commands = [
 		{ name: 'init-vs', cb: initVs },
 		{ name: 'init', cb: tt.init },