+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `tt` restart and status to the command palette.
 - Added `tt` install Tarantool CE to the command palette.
 - Added plugin icon.
+- Now the extension performs startup checks whether there are Tarantool annotations when it starts.
 
 ### Changed